[edt] Rename Float3Item.h/cpp to FloatXItem.h/cpp.
